Reduce SCSS: rewrite Swage as CSS (#8200)
* Reduce unused SCSS

* Rename Swage .scss to .css

* Convert from SCSS to native CSS
Contribute to https://github.com/FreshRSS/FreshRSS/issues/7611

Taking advantage of:
* https://developer.mozilla.org/en-US/docs/Web/CSS/Guides/Nesting
* https://developer.mozilla.org/en-US/docs/Web/CSS/Reference/Selectors/Nesting_selector
* https://developer.mozilla.org/en-US/docs/Web/CSS/Guides/Colors/Using_relative_colors#using_math_functions

This was mostly to get started and get a better idea of the difficulties.

I am happy to see that the differences between the original SCSS and the native CSS are limited.
The relevant diff to look at is: https://github.com/FreshRSS/FreshRSS/pull/8200/commits/dd1bc7c663cf7afdde7af3161ddc213d360247ef

The main remaining issue is that the *CSS Color Module Level 4, Relative Color* is not mainstream yet, so a small compatibility section is needed with precomputed colour values.